Now showing locations in the reminders list
Added a locationDescription field to Location to have something to
show in the list other than coordinates.
Reminder now has a method getLocationDescription() to get a display
string for that ocation. It returns locationDescription for manual
reminders and the poi for automatic reminders.
Also added a default constructor to Location for sugar orm.
Location.save() needs to be called explicitly when persisting a
Manualreminder.